CopyWithTooltip component
